+++ Remotes/IDigiMarket.cs
@@ -48,7 +48,7 @@ namespace Remotes
         void AddFundsDirect(string user, decimal balance, int diginoteCount);
         void OrdersSnapshot(List<PurchaseOrder> purchaseOrders, List<SalesOrder> salesOrders);
         void SetFundsDirect(string user, decimal balance);
-        void UpdateDiginotesDirect(string user, int diginotes, int value);
+        void UpdateDiginotesDirect(string user, int diginotes, decimal value);
         void ApplyTransaction(string buyer, string seller, int diginotes, decimal cost, DateTime date);
 
         #endregion
